// Jobs.Frontend/src/app/core/api/model/companyResponsePaginatedListResult.ts
/**
 * Jobs API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CompanyResponse } from './companyResponse';


/**
 * Represents a paginated result set.
 */
export interface CompanyResponsePaginatedListResult { 
    items?: Array<CompanyResponse> | null;
    totalCount?: number;
    pageNumber?: number;
    pageSize?: number;
    totalPages?: number;
}